+++ test/clj/rems/api/test_catalogue_items.clj
@@ -143,7 +143,7 @@
       (is (response-is-unauthorized? response))
       (is (str/includes? body "Invalid anti-forgery token"))))
   (testing "create with wrong API-Key"
-    (is (= "unauthorized"
+    (is (= "Invalid anti-forgery token"
            (-> (request :post (str "/api/catalogue-items/create"))
                (assoc-in [:headers "x-rems-api-key"] "invalid-api-key")
                (json-body {:form 1
@@ -161,7 +161,7 @@
       (is (response-is-unauthorized? response))
       (is (str/includes? body "Invalid anti-forgery token"))))
   (testing "edit with wrong API-Key"
-    (is (= "unauthorized"
+    (is (= "Invalid anti-forgery token"
            (-> (request :put (str "/api/catalogue-items/edit"))
                (assoc-in [:headers "x-rems-api-key"] "invalid-api-key")
                (json-body {:id 1